# server/swagger.json
{"swagger":"2.0","info":{"description":"wallet API documentation","version":"0.0.1","title":"wallet API"},"basePath":"/","tags":[],"schemes":["http"],"securityDefinitions":{"bearer":{"type":"apiKey","name":"Authorization","in":"header"}},"paths":{"/api/authenticate":{"post":{"summary":"Authorization api retrieving token","parameters":[{"name":"UserLoginDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/UserLoginDTO"}}],"responses":{"201":{"description":"Authorized"}},"tags":["user-jwt-controller"],"produces":["application/json"],"consumes":["application/json"]},"get":{"summary":"Check if the user is authenticated","responses":{"200":{"description":"login authenticated"}},"tags":["account-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/users":{"get":{"summary":"Get the list of users","responses":{"200":{"description":"List all users records","schema":{"$ref":"#/definitions/UserDTO"}}},"tags":["public-user-controller"],"produces":["application/json"],"consumes":["application/json"]}},"/api/authorities":{"get":{"summary":"Get the list of user roles","responses":{"200":{"description":"List all user roles","schema":{"type":"array","items":{"type":"string"}}}},"tags":["public-user-controller"],"produces":["application/json"],"consumes":["application/json"]}},"/api/verify/phone_number/init":{"post":{"summary":"Initiate verify Phone number by OTP","description":"Initiate the process if verifying a phone number by Sending an OTP to the user's phone number provided in the request","parameters":[{"name":"InitiateVerifyPhoneNumberDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/InitiateVerifyPhoneNumberDTO"}}],"responses":{"201":{"description":"OTP verification token, that must be sent in the payload when verifying the OTP"}},"tags":["account-resource"],"produces":["application/json"],"consumes":["application/json"]}},"/api/verify/phone_number/finish":{"post":{"summary":"Verify Phone number by OTP","description":"Verify a phone number validating the otp provided in the request.","parameters":[{"name":"VerifyPhoneNumberDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/VerifyPhoneNumberDTO"}}],"responses":{"201":{"description":"OTP verification token, that must be sent in the payload when verifying the OTP"}},"tags":["account-resource"],"produces":["application/json"],"consumes":["application/json"]}},"/api/register":{"post":{"summary":"Register user","parameters":[],"responses":{"201":{"description":"Registered user","schema":{"$ref":"#/definitions/UserDTO"}}},"tags":["account-resource"],"produces":["application/json"],"consumes":["application/json"]}},"/api/activate":{"get":{"summary":"Activate an account","parameters":[],"responses":{"200":{"description":"activated"}},"tags":["account-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/account":{"get":{"summary":"Get the current user.","responses":{"200":{"description":"user retrieved"}},"tags":["account-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"post":{"summary":"Update the current user information","parameters":[{"name":"UserDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/UserDTO"}}],"responses":{"201":{"description":"user info updated","schema":{"$ref":"#/definitions/UserDTO"}}},"tags":["account-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/account/change-password":{"post":{"summary":"Change current password","parameters":[{"name":"PasswordChangeDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/PasswordChangeDTO"}}],"responses":{"201":{"description":"user password changed","schema":{"$ref":"#/definitions/PasswordChangeDTO"}}},"tags":["account-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/account/reset-password/init":{"post":{"summary":"Send an email to reset the password of the user","parameters":[{"name":"PasswordResetDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/PasswordResetDTO"}}],"responses":{"201":{"description":"mail to reset password sent","schema":{"type":"string"}}},"tags":["account-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/account/reset-password/finish":{"post":{"summary":"Finish to reset the password of the user","parameters":[{"type":"string","name":"String","required":true,"in":"body"}],"responses":{"201":{"description":"password reset","schema":{"type":"string"}}},"tags":["account-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/admin/users":{"get":{"summary":"Get the list of users","responses":{"200":{"description":"List all users","schema":{"$ref":"#/definitions/UserDTO"}}},"tags":["user-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"post":{"summary":"Create user","parameters":[{"name":"UserDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/UserDTO"}}],"responses":{"201":{"description":"The record has been successfully created.","schema":{"$ref":"#/definitions/UserDTO"}},"403":{"description":"Forbidden."}},"tags":["user-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update user","parameters":[{"name":"UserDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/UserDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/UserDTO"}}},"tags":["user-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/admin/users/{login}":{"get":{"summary":"Get user","parameters":[{"type":"string","name":"login","required":true,"in":"path"}],"responses":{"200":{"description":"The found record","schema":{"$ref":"#/definitions/UserDTO"}}},"tags":["user-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"delete":{"summary":"Delete user","parameters":[{"type":"string","name":"login","required":true,"in":"path"}],"responses":{"204":{"description":"The record has been successfully deleted.","schema":{"$ref":"#/definitions/UserDTO"}}},"tags":["user-resource"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/customers":{"get":{"responses":{"200":{"description":"List all records","schema":{"$ref":"#/definitions/CustomerDTO"}}},"tags":["customers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"post":{"summary":"Create customerEntity","parameters":[{"name":"CustomerDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/CustomerDTO"}}],"responses":{"201":{"description":"The record has been successfully created.","schema":{"$ref":"#/definitions/CustomerDTO"}},"403":{"description":"Forbidden."}},"tags":["customers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update customerEntity","parameters":[{"name":"CustomerDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/CustomerDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/CustomerDTO"}}},"tags":["customers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/customers/{id}":{"get":{"parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":"The found record","schema":{"$ref":"#/definitions/CustomerDTO"}}},"tags":["customers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update customerEntity with id","parameters":[{"name":"CustomerDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/CustomerDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/CustomerDTO"}}},"tags":["customers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"delete":{"summary":"Delete customerEntity","parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"204":{"description":"The record has been successfully deleted."}},"tags":["customers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/customers/{id}/bvn/add":{"post":{"parameters":[{"name":"BvnDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/BvnDTO"}},{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":"Add customer's BVN to their profile"}},"tags":["customers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/customers/{id}/next_of_kin/add":{"post":{"parameters":[{"name":"NextOfKinDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/NextOfKinDTO"}},{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":"Add customer's next of kin to their profile"}},"tags":["customers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/wallets":{"get":{"responses":{"200":{"description":"List all records","schema":{"$ref":"#/definitions/WalletDTO"}}},"tags":["wallets"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"post":{"summary":"Create walletEntity","parameters":[{"name":"WalletDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/WalletDTO"}}],"responses":{"201":{"description":"The record has been successfully created.","schema":{"$ref":"#/definitions/WalletDTO"}},"403":{"description":"Forbidden."}},"tags":["wallets"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update walletEntity","parameters":[{"name":"WalletDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/WalletDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/WalletDTO"}}},"tags":["wallets"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/wallets/{id}":{"get":{"parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":"The found record","schema":{"$ref":"#/definitions/WalletDTO"}}},"tags":["wallets"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update walletEntity with id","parameters":[{"name":"WalletDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/WalletDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/WalletDTO"}}},"tags":["wallets"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"delete":{"summary":"Delete walletEntity","parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"204":{"description":"The record has been successfully deleted."}},"tags":["wallets"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/wallets/{id}/transfer":{"post":{"summary":"Transfer funds to another wallet in the system.","parameters":[{"name":"SendMoneyDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/SendMoneyDTO"}},{"type":"string","name":"id","required":true,"in":"path"}],"responses":{"201":{"description":"Transfer Successful.","schema":{"$ref":"#/definitions/WalletDTO"}},"403":{"description":"Forbidden."}},"tags":["wallets"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/wallets/{id}/fund-wallet":{"post":{"summary":"Funding your wallet from an external source.","parameters":[{"name":"SendMoneyDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/SendMoneyDTO"}},{"type":"string","name":"id","required":true,"in":"path"}],"responses":{"201":{"description":"Deposit Successful.","schema":{"$ref":"#/definitions/WalletDTO"}},"403":{"description":"Forbidden."}},"tags":["wallets"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/journal-lines":{"get":{"responses":{"200":{"description":"List all records","schema":{"$ref":"#/definitions/JournalLineDTO"}}},"tags":["journal-lines"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"post":{"summary":"Create journalLineEntity","parameters":[{"name":"JournalLineDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/JournalLineDTO"}}],"responses":{"201":{"description":"The record has been successfully created.","schema":{"$ref":"#/definitions/JournalLineDTO"}},"403":{"description":"Forbidden."}},"tags":["journal-lines"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update journalLineEntity","parameters":[{"name":"JournalLineDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/JournalLineDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/JournalLineDTO"}}},"tags":["journal-lines"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/journal-lines/{id}":{"get":{"parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":"The found record","schema":{"$ref":"#/definitions/JournalLineDTO"}}},"tags":["journal-lines"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update journalLineEntity with id","parameters":[{"name":"JournalLineDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/JournalLineDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/JournalLineDTO"}}},"tags":["journal-lines"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"delete":{"summary":"Delete journalLineEntity","parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"204":{"description":"The record has been successfully deleted."}},"tags":["journal-lines"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/journals":{"get":{"responses":{"200":{"description":"List all records","schema":{"$ref":"#/definitions/JournalDTO"}}},"tags":["journals"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"post":{"summary":"Create journalEntity","parameters":[{"name":"JournalDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/JournalDTO"}}],"responses":{"201":{"description":"The record has been successfully created.","schema":{"$ref":"#/definitions/JournalDTO"}},"403":{"description":"Forbidden."}},"tags":["journals"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update journalEntity","parameters":[{"name":"JournalDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/JournalDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/JournalDTO"}}},"tags":["journals"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/journals/{id}":{"get":{"parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":"The found record","schema":{"$ref":"#/definitions/JournalDTO"}}},"tags":["journals"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update journalEntity with id","parameters":[{"name":"JournalDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/JournalDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/JournalDTO"}}},"tags":["journals"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"delete":{"summary":"Delete journalEntity","parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"204":{"description":"The record has been successfully deleted."}},"tags":["journals"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/app-transacts":{"get":{"responses":{"200":{"description":"List all records","schema":{"$ref":"#/definitions/AppTransactDTO"}}},"tags":["app-transacts"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"post":{"summary":"Create appTransactEntity","parameters":[{"name":"AppTransactDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/AppTransactDTO"}}],"responses":{"201":{"description":"The record has been successfully created.","schema":{"$ref":"#/definitions/AppTransactDTO"}},"403":{"description":"Forbidden."}},"tags":["app-transacts"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update appTransactEntity","parameters":[{"name":"AppTransactDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/AppTransactDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/AppTransactDTO"}}},"tags":["app-transacts"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/app-transacts/{id}":{"get":{"parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":"The found record","schema":{"$ref":"#/definitions/AppTransactDTO"}}},"tags":["app-transacts"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update appTransactEntity with id","parameters":[{"name":"AppTransactDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/AppTransactDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/AppTransactDTO"}}},"tags":["app-transacts"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"delete":{"summary":"Delete appTransactEntity","parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"204":{"description":"The record has been successfully deleted."}},"tags":["app-transacts"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/billers":{"get":{"responses":{"200":{"description":"List all records","schema":{"$ref":"#/definitions/BillerDTO"}}},"tags":["billers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"post":{"summary":"Create billerEntity","parameters":[{"name":"BillerDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/BillerDTO"}}],"responses":{"201":{"description":"The record has been successfully created.","schema":{"$ref":"#/definitions/BillerDTO"}},"403":{"description":"Forbidden."}},"tags":["billers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update billerEntity","parameters":[{"name":"BillerDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/BillerDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/BillerDTO"}}},"tags":["billers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/api/billers/{id}":{"get":{"parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"200":{"description":"The found record","schema":{"$ref":"#/definitions/BillerDTO"}}},"tags":["billers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"Update billerEntity with id","parameters":[{"name":"BillerDTO","required":true,"in":"body","schema":{"$ref":"#/definitions/BillerDTO"}}],"responses":{"200":{"description":"The record has been successfully updated.","schema":{"$ref":"#/definitions/BillerDTO"}}},"tags":["billers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"delete":{"summary":"Delete billerEntity","parameters":[{"type":"number","name":"id","required":true,"in":"path"}],"responses":{"204":{"description":"The record has been successfully deleted."}},"tags":["billers"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}}},"definitions":{"UserLoginDTO":{"type":"object","properties":{"password":{"type":"string","description":"User password"},"rememberMe":{"type":"boolean","description":"User remember login"},"username":{"type":"string","description":"User login name"}},"required":["password","username"]},"UserDTO":{"type":"object","properties":{"login":{"type":"string","uniqueItems":true,"example":"myuser","description":"User login"},"firstName":{"type":"string","example":"MyUser","description":"User first name"},"lastName":{"type":"string","example":"MyUser","description":"User last name"},"email":{"type":"string","example":"myuser@localhost.it","description":"User email"},"activated":{"type":"boolean","example":"true","description":"User activation"},"langKey":{"type":"string","example":"en","description":"User language"},"authorities":{"type":"array","description":"Array of permissions","items":{"type":"string","enum":["ROLE_USER","ROLE_ADMIN","ROLE_ANONYMOUS"]}},"password":{"type":"string","example":"myuser","description":"User password"},"imageUrl":{"type":"string","example":"http://my-image-url","description":"Image url"}},"required":["login","email","password"]},"InitiateVerifyPhoneNumberDTO":{"type":"object","properties":{"phoneNumber":{"type":"string","description":"The phone number to verify. The country code must be included, in the format \"[+county code][phone number] Eg +2340XXXXXXXXX\"."}},"required":["phoneNumber"]},"VerifyPhoneNumberDTO":{"type":"object","properties":{"otp":{"type":"string","description":"The otp value sent to a user via sms"},"otpKey":{"type":"string","description":"A string representing an encrypted set of information about the otp received from the backend when the otp from requested"}},"required":["otp","otpKey"]},"PasswordChangeDTO":{"type":"object","properties":{"currentPassword":{"type":"string","description":"Current password"},"newPassword":{"type":"string","description":"New password"}},"required":["currentPassword","newPassword"]},"PasswordResetDTO":{"type":"object","properties":{"email":{"type":"string","description":"User's email"}},"required":["email"]},"CustomerDTO":{"type":"object","properties":{"customerID":{"type":"string","description":"customerID field"},"phoneNumber":{"type":"string","description":"phoneNumber field"},"nationality":{"type":"string","description":"nationality field"},"address":{"type":"string","description":"address field"},"bvn":{"type":"number","description":"bvn field"},"kycLevel":{"type":"number","description":"KYC level field"},"dateOfBirth":{"type":"object","description":"dateOfBirth field"},"status":{"type":"string","description":"status field","enum":["Active","Inactive"],"default":"Active"},"user":{"title":"UserDTO","allOf":[{"$ref":"#/definitions/UserDTO"},{"description":"user relationship"}]}},"required":["customerID","phoneNumber","user"]},"BvnDTO":{"type":"object","properties":{"bvn":{"type":"number","description":"The user's Bank Verification Number, must an 11 digit number."}},"required":["bvn"]},"NextOfKinDTO":{"type":"object","properties":{"firstName":{"type":"string","example":"John","description":"Next of Kin first name"},"lastName":{"type":"string","example":"Doe","description":"Next of Kin last name"},"email":{"type":"string","example":"myuser@localhost.it","description":"Next of Kin email"},"phoneNumber":{"type":"string","description":"Next of Kin phoneNumber"},"nationality":{"type":"string","description":"Next of Kin nationality"},"address":{"type":"string","description":"Next of Kin address"},"relationship":{"type":"string","description":"User relationship with next of kin","enum":["Father","Mother","Sister","Brother","Friend"]},"customer":{"title":"CustomerDTO","allOf":[{"$ref":"#/definitions/CustomerDTO"},{"description":"customer relationship"}]}},"required":["firstName","lastName","email","phoneNumber","nationality","address","relationship","customer"]},"WalletDTO":{"type":"object","properties":{"walletID":{"type":"string","description":"walletID field"},"nuban":{"type":"string","description":"nuban field"},"accountNumber":{"type":"string","description":"accountNumber field"},"balance":{"type":"number","description":"balance field"},"expiryDate":{"type":"object","description":"expiryDate field"},"type":{"type":"string","enum":["BonusWallet","CustomerWallet","BillerWallet","SystemWallet"],"description":"type enum field"},"status":{"type":"string","enum":["Inactive","Active","Blocked","Suspended"],"description":"status enum field"},"customer":{"title":"CustomerDTO","allOf":[{"$ref":"#/definitions/CustomerDTO"},{"description":"customer relationship"}]}},"required":["customer"]},"SendMoneyDTO":{"type":"object","properties":{"amount":{"type":"number","description":"amount of money to be sent"},"originWalletID":{"type":"string","description":"wallet ID of the wallet money is to be sent from"},"destinationWalletID":{"type":"string","description":"wallet ID of the wallet money is to be sent to"},"transactionType":{"type":"string","description":"accompanying narration"},"narration":{"type":"string","description":"accompanying narration"}},"required":["amount","originWalletID","destinationWalletID","transactionType"]},"JournalDTO":{"type":"object","properties":{"journalID":{"type":"string","description":"journalID field"},"dateOfEntry":{"type":"object","description":"dateOfEntry field"},"description":{"type":"string","description":"description field"},"journalLines":{"type":"array","description":"journalLines relationship","items":{"type":"array"}}},"required":["journalLines"]},"JournalLineDTO":{"type":"object","properties":{"amount":{"type":"number","description":"amount field"},"type":{"type":"string","enum":["WalletTransfer","BankTransfer","WalletFunding","BankFunding","BillPayment"],"description":"type enum field"},"narration":{"type":"string","description":"narration field"},"operation":{"type":"string","enum":["Debit","Credit"],"description":"operation enum field"},"journal":{"title":"JournalDTO","allOf":[{"$ref":"#/definitions/JournalDTO"},{"description":"journal relationship"}]},"wallet":{"title":"WalletDTO","allOf":[{"$ref":"#/definitions/WalletDTO"},{"description":"wallet relationship"}]}},"required":["journal","wallet"]},"AppTransactDTO":{"type":"object","properties":{"type":{"type":"string","enum":["WalletTransfer","BankTransfer","WalletFunding","BankFunding","BillPayment"],"description":"type enum field"},"transactionDate":{"type":"object","description":"transactionDate field"},"transactionRef":{"type":"string","description":"transactionRef field"},"description":{"type":"string","description":"description field"},"customer":{"title":"CustomerDTO","allOf":[{"$ref":"#/definitions/CustomerDTO"},{"description":"customer relationship"}]}},"required":["customer"]},"BillerDTO":{"type":"object","properties":{"name":{"type":"string","description":"name field"},"description":{"type":"string","description":"description field"},"wallet":{"title":"WalletDTO","allOf":[{"$ref":"#/definitions/WalletDTO"},{"description":"wallet relationship"}]}},"required":["wallet"]}}}